config: allow using custom PAM configs
In the event several greeters are installed, each needing a specific PAM
config, it could be useful to have each greeter having its own config
file (which is already possible), each one indicating the PAM config to
be used with this greeter (which is not possible for now).

This commit adds a new (optional) "service" key to the "general",
"default_session" and "initial_session" sections so users and/or
distributions can have more flexibility regarding PAM config management.

In case this option is missing from the config file, we fall back to the
default "greetd" and "greetd-greeter" config names.

Signed-off-by: Arnaud Ferraris <arnaud.ferraris@collabora.com>