greetd: Use pointer::is_null instead of cast
While dealing with the large number of casts needed in the PAM ffi code,
the return value of calloc was accidentally cast to a u8 for the error
check, meaning it erroneously thought calloc had failed if the least
significant byte was 0 - which it often wasn't, meaning that users did
not discover the issue.

Avoid casting and just use pointer::is_null.